refactor: 重构菜单规格

+++ dal/model/ent/order/where.go
@@ -11,47 +11,47 @@ import (
 )
 
 // ID filters vertices based on their ID field.
-func ID(id int) predicate.Order {
+func ID(id uint64) predicate.Order {
 	return predicate.Order(sql.FieldEQ(FieldID, id))
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id int) predicate.Order {
+func IDEQ(id uint64) predicate.Order {
 	return predicate.Order(sql.FieldEQ(FieldID, id))
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id int) predicate.Order {
+func IDNEQ(id uint64) predicate.Order {
 	return predicate.Order(sql.FieldNEQ(FieldID, id))
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...int) predicate.Order {
+func IDIn(ids ...uint64) predicate.Order {
 	return predicate.Order(sql.FieldIn(FieldID, ids...))
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...int) predicate.Order {
+func IDNotIn(ids ...uint64) predicate.Order {
 	return predicate.Order(sql.FieldNotIn(FieldID, ids...))
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id int) predicate.Order {
+func IDGT(id uint64) predicate.Order {
 	return predicate.Order(sql.FieldGT(FieldID, id))
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id int) predicate.Order {
+func IDGTE(id uint64) predicate.Order {
 	return predicate.Order(sql.FieldGTE(FieldID, id))
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id int) predicate.Order {
+func IDLT(id uint64) predicate.Order {
 	return predicate.Order(sql.FieldLT(FieldID, id))
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id int) predicate.Order {
+func IDLTE(id uint64) predicate.Order {
 	return predicate.Order(sql.FieldLTE(FieldID, id))
 }
