+++ dal/model/ent/orderitem/where.go
@@ -11,47 +11,47 @@ import (
 )
 
 // ID filters vertices based on their ID field.
-func ID(id int) predicate.OrderItem {
+func ID(id uint64) predicate.OrderItem {
 	return predicate.OrderItem(sql.FieldEQ(FieldID, id))
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id int) predicate.OrderItem {
+func IDEQ(id uint64) predicate.OrderItem {
 	return predicate.OrderItem(sql.FieldEQ(FieldID, id))
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id int) predicate.OrderItem {
+func IDNEQ(id uint64) predicate.OrderItem {
 	return predicate.OrderItem(sql.FieldNEQ(FieldID, id))
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...int) predicate.OrderItem {
+func IDIn(ids ...uint64) predicate.OrderItem {
 	return predicate.OrderItem(sql.FieldIn(FieldID, ids...))
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...int) predicate.OrderItem {
+func IDNotIn(ids ...uint64) predicate.OrderItem {
 	return predicate.OrderItem(sql.FieldNotIn(FieldID, ids...))
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id int) predicate.OrderItem {
+func IDGT(id uint64) predicate.OrderItem {
 	return predicate.OrderItem(sql.FieldGT(FieldID, id))
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id int) predicate.OrderItem {
+func IDGTE(id uint64) predicate.OrderItem {
 	return predicate.OrderItem(sql.FieldGTE(FieldID, id))
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id int) predicate.OrderItem {
+func IDLT(id uint64) predicate.OrderItem {
 	return predicate.OrderItem(sql.FieldLT(FieldID, id))
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id int) predicate.OrderItem {
+func IDLTE(id uint64) predicate.OrderItem {
 	return predicate.OrderItem(sql.FieldLTE(FieldID, id))
 }
