+++ dal/model/ent/table/where.go
@@ -11,47 +11,47 @@ import (
 )
 
 // ID filters vertices based on their ID field.
-func ID(id int) predicate.Table {
+func ID(id uint64) predicate.Table {
 	return predicate.Table(sql.FieldEQ(FieldID, id))
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id int) predicate.Table {
+func IDEQ(id uint64) predicate.Table {
 	return predicate.Table(sql.FieldEQ(FieldID, id))
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id int) predicate.Table {
+func IDNEQ(id uint64) predicate.Table {
 	return predicate.Table(sql.FieldNEQ(FieldID, id))
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...int) predicate.Table {
+func IDIn(ids ...uint64) predicate.Table {
 	return predicate.Table(sql.FieldIn(FieldID, ids...))
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...int) predicate.Table {
+func IDNotIn(ids ...uint64) predicate.Table {
 	return predicate.Table(sql.FieldNotIn(FieldID, ids...))
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id int) predicate.Table {
+func IDGT(id uint64) predicate.Table {
 	return predicate.Table(sql.FieldGT(FieldID, id))
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id int) predicate.Table {
+func IDGTE(id uint64) predicate.Table {
 	return predicate.Table(sql.FieldGTE(FieldID, id))
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id int) predicate.Table {
+func IDLT(id uint64) predicate.Table {
 	return predicate.Table(sql.FieldLT(FieldID, id))
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id int) predicate.Table {
+func IDLTE(id uint64) predicate.Table {
 	return predicate.Table(sql.FieldLTE(FieldID, id))
 }
