+++ dal/model/ent/tablecategory/where.go
@@ -11,47 +11,47 @@ import (
 )
 
 // ID filters vertices based on their ID field.
-func ID(id int) predicate.TableCategory {
+func ID(id uint64) predicate.TableCategory {
 	return predicate.TableCategory(sql.FieldEQ(FieldID, id))
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id int) predicate.TableCategory {
+func IDEQ(id uint64) predicate.TableCategory {
 	return predicate.TableCategory(sql.FieldEQ(FieldID, id))
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id int) predicate.TableCategory {
+func IDNEQ(id uint64) predicate.TableCategory {
 	return predicate.TableCategory(sql.FieldNEQ(FieldID, id))
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...int) predicate.TableCategory {
+func IDIn(ids ...uint64) predicate.TableCategory {
 	return predicate.TableCategory(sql.FieldIn(FieldID, ids...))
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...int) predicate.TableCategory {
+func IDNotIn(ids ...uint64) predicate.TableCategory {
 	return predicate.TableCategory(sql.FieldNotIn(FieldID, ids...))
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id int) predicate.TableCategory {
+func IDGT(id uint64) predicate.TableCategory {
 	return predicate.TableCategory(sql.FieldGT(FieldID, id))
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id int) predicate.TableCategory {
+func IDGTE(id uint64) predicate.TableCategory {
 	return predicate.TableCategory(sql.FieldGTE(FieldID, id))
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id int) predicate.TableCategory {
+func IDLT(id uint64) predicate.TableCategory {
 	return predicate.TableCategory(sql.FieldLT(FieldID, id))
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id int) predicate.TableCategory {
+func IDLTE(id uint64) predicate.TableCategory {
 	return predicate.TableCategory(sql.FieldLTE(FieldID, id))
 }
